# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for NewsApi Application",
        "title": "News API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/news": {
            "get": {
                "description": "get news",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Get all news",
                "parameters": [
                    {
                        "type": "number",
                        "format": "number",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "string",
                        "description": "fromDate",
                        "name": "fromDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "toDate",
                        "description": "toDate",
                        "name": "toDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/router.getNewsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/router.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/router.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/router.errorResponse"
                        }
                    }
                }
            }
        },
        "/news/{id}": {
            "get": {
                "description": "get news by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Get news By ID",
                "operationId": "get-news-by-id",
                "parameters": [
                    {
                        "type": "number",
                        "description": "ID list",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.NewsList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/router.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/router.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/router.errorResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/router.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.NewsList": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "TEXT",
                    "type": "string"
                },
                "id": {
                    "description": "Поле id как SERIAL",
                    "type": "string"
                },
                "link": {
                    "description": "VARCHAR(255) UNIQUE",
                    "type": "string"
                },
                "publishedAt": {
                    "description": "TIMESTAMP",
                    "type": "string"
                },
                "title": {
                    "description": "VARCHAR(255)",
                    "type": "string"
                }
            }
        },
        "router.errorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "router.getNewsResponse": {
            "type": "object",
            "properties": {
                "currentPage": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.NewsList"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        }
    }
}